+++ src/symbol.c
@@ -1,8 +1,9 @@
 
+#include "clue.h"
 #include "symbol.h"
 #include "util.h"
 
-Symbol* newSymbol(const char* name, unsigned int precedence, bool reserved) {
+Symbol* newSymbol(const char* name, u64 precedence, bool reserved) {
     Symbol* symbol = pMalloc(sizeof (Symbol));
 
     symbol->name = name;
